% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPMD.R
\name{FPMD}
\alias{FPMD}
\title{Functional Processes with Multiple Discontinuities}
\usage{
FPMD(
  Ly,
  Lt,
  wi = NULL,
  Wtype = c("OBS", "SUBJ", "MIX", "OPT"),
  zeta = NULL,
  cutoff = max,
  alpha = 0.05,
  bw.seq = NULL,
  nRegGrid = 101,
  NbGrid = 101,
  kernel = "epan",
  npoly = 1,
  nder = 0,
  kFolds = 5,
  refined = FALSE,
  individualCurveJump = FALSE,
  M_max
)
}
\arguments{
\item{Ly}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{Lt}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{wi}{A vector of weight for each observation.}

\item{Wtype}{Weight structure specified for estimation. It works if \code{wi = NULL}.}

\item{zeta}{Cut-off threshold for change point detection. The default is \code{NULL}.}

\item{cutoff}{The method for construct the cut-off statistics. Default is \code{max}. For the case
observations are much larger than subjections, we suggest to use the \code{min}.}

\item{alpha}{The confidence level for the cut-off threshold. The default is
\code{alpha = 0.05}, and for heavy tailed process, we suggest a higher level \code{0.1}.}

\item{bw.seq}{The selected (or user specified) bandwidth for cross validation. The default is \code{NULL}.}

\item{nRegGrid}{The number of support points in each direction of covariance surface; numeric - default: 101.}

\item{NbGrid}{The number of support points in the jump points detection; numeric - default: 101.}

\item{kernel}{Smoothing kernel choice, common for mu and covariance; "rect", "epan", "quar" - default: "epan".}

\item{npoly}{The degree of polynomial. Default is 1 for local linear smoothing.}

\item{nder}{The order of derivative. Default is 0 for local linear smoothing, and should smaller than npoly.}

\item{kFolds}{Number of folds for cross validation - default: 5.}

\item{refined}{A refined stage for jump size estimation after the jump locations are detected; logical - default: TRUE}

\item{individualCurveJump}{Jump detection for each curve by the BIC proposed in Xia and Qiu, 2016.}

\item{M_max}{The maximum number of jumps for individual curve when using the BIC.}
}
\value{
A list containing the following fields: If the change points are detected, the function
output the followings
\item{mu_jumptime}{The detected change point locations. If no change point detected, \code{mu_jumptime=0}}
\item{mu_jumpsize}{The estimated jump sizes. If no change point detected, \code{mu_jumpsize=0}}
\item{mu_jumpsize_h_tau}{The detected change point locations. If no change point detected, \code{mu_jumptime=0}}
\item{zeta}{The cut-off threshold for change point detection.}
\item{mu}{For no change point case. A vector of length obsGrid containing the mean function estimate.}
\item{muWork}{For change point case. A vector of length obsGrid containing the mean function estimate.}
\item{wi}{The weight for each observation.}
\item{mi}{Number of observations for each subject.}
\item{timings}{A vector with execution times for the basic parts of the FMbreaks call.}
\item{timings_jump}{A vector with execution times for the basic parts of jump detection.}
\item{workGrid}{A vector of length nWorkGrid.}
\item{regGrid}{A vector of length regGrid.}
\item{sigma2}{Variance for measure error.}
\item{smoothCov}{A nWorkGrid by nWorkGrid matrix of the smoothed covariance surface.}
\item{fittedCov}{A nWorkGrid by nWorkGrid matrix of the fitted covariance surface, which is guaranteed to be non-negative definite.}
\item{optns}{A list of actually used options.}
\item{h_tau}{The selected (or user specified) bandwidth for jump locations estimate.}
\item{h_d}{The selected (or user specified) bandwidth for jump sizes estiamte.}
\item{LX}{The fitted individual curves.}
\item{indJump}{The jump detection for each individual curve.}
}
\description{
Multiple Thresholds detection in mean function for dense or sparse functional data.
The bandwidths are selected by a k-folds cross validation.
}
\examples{

\dontrun{
# setting 1
## 3 change points
tau = c(.25,.5,.75) # jump locations
disc = c(.5,-.4,.4) # jump sizes
smoothmean=function(x) x^2+sin(2*pi*x) + cos(2*pi*x)
jump = function(x) sum((x>=tau)*disc) # jump function
mu_fun= function(t){smoothmean(t)+ sapply(t, jump)}
n = 400
Q = 20

# set.seed(123)
## generate gaussian process
## "cos", "sin", "fourier", "legendre01", "poly"
data = MakeFPMBData(n = n, Q = Q, proc= "gaussian", muFun = mu_fun, K = 3,
                    lambda = 1/(seq_len(3)+1)^2, sigma = 0.2, basisType = "fourier")
Lt = data$Lt
Ly = data$Ly
bw.seq = seq(0.01, 0.13, by = 0.02)
system.time(resCP <- FPMD(Ly = Ly, Lt = Lt, wi = NULL, Wtype = "MIX", zeta = NULL,
                              bw.seq = 0.06, NbGrid = 101, kFolds = 5, refined = TRUE,
                              individualCurveJump = FALSE, nRegGrid = 101,
                              npoly = 1, nder = 0, alpha = 0.05, cutoff = max))
h_tau = resCP$h_tau
h_d = resCP$h_d
zeta = resCP$zeta
mu_jumptime = resCP$mu_jumptime
mu_jumpsize = resCP$mu_jumpsize

## pointwise confidence band
PCBplot(res = resCP)
lines(resCP$obsGrid, mu_fun(resCP$obsGrid), type="l", col = 'blue' )

}


}
\references{
\cite{Li, J., Li, Y., and Hsing, T. (2021). " On Functional Processes with Multiple Discontinuities".}
}
