% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_GenData.R
\name{MakeFPMBData}
\alias{MakeFPMBData}
\title{Create a Dense Functional Data sample for a Gaussian process}
\usage{
MakeFPMBData(
  n,
  Q,
  muFun,
  rdist = runif,
  K,
  lambda,
  sigma,
  basisType = "cos",
  proc = c("gaussian", "t")
)
}
\arguments{
\item{n}{number of samples to generate.}

\item{Q}{an integral. The observations for per sample are following the poisson distribution with mean Q.}

\item{muFun}{a function that takes a vector input and output a vector of the corresponding mean (default: zero function).}

\item{rdist}{a sampler for generating the random design time points within [0, 1].}

\item{K}{scalar specifying the number of basis to be used (default: 3).}

\item{lambda}{vector of size K specifying the variance of each components (default: rep(1,K)).}

\item{sigma}{The standard deviation of the Gaussian noise added to each observation points.}

\item{basisType}{string specifying the basis type used; possible options are: 'sin', 'cos' and 'fourier' (default: 'cos') (See code of 'CreateBasis' for implementation details.)}

\item{proc}{The type of random process, 'gaussian' or 't' process.}
}
\value{
A list containing the generated data with \code{List} type:
\item{Ly}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}
\item{Lt}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}
\item{mi}{Number of observations for each subject.}
}
\description{
For a Gaussian or t process, create a dense or sparse functional data sample of size n over a [0,1] support.
}
