% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalTest.R
\name{MDTest}
\alias{MDTest}
\title{A global test for jump points in mean function}
\usage{
MDTest(
  bw,
  alpha = 0.05,
  Lt,
  Ly,
  nRegGrid = 101,
  Wtype = c("OBS", "SUBJ", "MIX", "OPT")
)
}
\arguments{
\item{bw}{The selected (or user specified) bandwidth.}

\item{alpha}{The confidence level for the cut-off threshold. The default is
\code{alpha = 0.05}, and for heavy tailed process, we suggest a higher level \code{0.1}.}

\item{Lt}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{Ly}{A list of \emph{n} vectors containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{nRegGrid}{The number of support points in each direction of covariance surface; numeric - default: 101.}

\item{Wtype}{Weight structure specified for estimation. It works if \code{wi = NULL}.}
}
\value{
\item{reject}{1 indicates that there exists discontinuities.}
}
\description{
A global test for jump points in mean function
}
