% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanBreaksFP.R
\name{MeanBreaksFP}
\alias{MeanBreaksFP}
\title{Multiple Thresholds detection in mean function}
\usage{
MeanBreaksFP(
  yin,
  xin,
  win,
  mi,
  xout,
  NbGrid,
  h_tau,
  h_d,
  zeta,
  npoly = 1,
  nder = 0,
  kernel,
  refined = FALSE
)
}
\arguments{
\item{yin}{A vector containing the observed values.}

\item{xin}{A vector containing the observation time points corresponding to yin.}

\item{win}{A list containing a vector of weight for each observations and the duplicated weight.}

\item{mi}{A vector containing the number of observations for each subject.}

\item{xout}{A vector containing the observation time points for mean estimation.}

\item{NbGrid}{The number of support points in the jump points detection; numeric - default: 101.}

\item{h_tau}{The selected (or user specified) bandwidth for jump locations estimate.}

\item{h_d}{The selected (or user specified) bandwidth for jump sizes estiamte.}

\item{zeta}{Cut-off threshold for change point detection.}

\item{npoly}{The degree of polynomial. Default is 1 for local linear smoothing.}

\item{nder}{The order of derivative. Default is 0 for local linear smoothing, and should smaller than npoly.}

\item{kernel}{Smoothing kernel choice, common for mu and covariance; "rect", "epan", "quar" - default: "epan".}

\item{refined}{A refined stage for jump size estimation after the jump locations are detected; logical - default: TRUE}
}
\value{
A list containing the following fields: If the change points are detected, the function
output the followings
\item{mu_jumptime}{The detected change point locations. If no change point detected, \code{mu_jumptime=0}}
\item{mu_jumpsize}{The estimated jump sizes. If no change point detected, \code{mu_jumpsize=0}}
\item{mu_jumpsize_h_tau}{The detected change point locations. If no change point detected, \code{mu_jumptime=0}}
\item{mu}{For no change point case. A vector of length obsGrid containing the mean function estimate.}
\item{muout}{For change point case. A vector of length xout containing the mean function estimate.}
\item{xout}{A vector containing the observation time points for mean estimation.}
\item{timings}{A vector with execution times for the basic parts of the FMbreaks call.}
\item{h_tau}{The selected (or user specified) bandwidth for jump locations estimate.}
\item{h_d}{The selected (or user specified) bandwidth for jump sizes estiamte.}
}
\description{
Multiple Thresholds detection in mean function for dense or sparse functional data.
}
\references{
\cite{Li, J., Li, Y., and Hsing, T. (2021). " On Functional Processes with Multiple Discontinuities".}
}
