% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{PCBplot}
\alias{PCBplot}
\title{Plot the estimated mean function and the corresponding pointwise confidence interval}
\usage{
PCBplot(res)
}
\arguments{
\item{res}{an FPMD class object returned by \code{FPMD()}.}
}
\description{
Plot the estimated mean function and the corresponding pointwise confidence interval
}
